refactor: Break up Tracer unit tests
