+++ test/unit/TracerSettle.js
@@ -30,7 +30,7 @@ const setupTests = deployments.createFixture(async () => {
     }
 })
 
-describe("Unit tests: settle", function () {
+describe("Unit tests: TracerPerpetualSwaps.sol settle", function () {
     context("when the account has no open positions", async () => {
         it("updates the last updated index and gas price but does not change the account balance", async () => {
             const { tracer, pricing, quoteToken, gasEthOracle } =
